Handling of some some modified deps for 'fixitude'.
Rewrote the main algorithm so that the full dependency tree is checked,
even if the current target seems up-to-date when looking at the
immediate dependencies. -- This allows for rebuilding of targets deeper
down in the dependency tree (if they were deleted) and discovery of
dependency changes that might require rebuilding of the target given by
the user.

This means that the core of fixitude is now what it should be, and
capable of handling any kind of dependency tree, and responding to
changes by only rebuilding a minimal amount of stuff. Yay! \o/

Also added a couple of more test to the 'fixitude-test' script, to test
for more stuff involving multiple dependencies.