Added test for fix $VERSION number.
The $VERSION number test uses the git repo to determine if 'fix' has
been updated since last commit, uses git tags to find out last
major/minor version numbers, and the number of commits that modified
'fix' to determine the patch number.

All that will be skipped if a git repo could not be found.