Output of '--man' now works on MacOS X.
The output is now plain text, instead of the fancy manpage format. This
is a workaround the fact that the plain 'man' (as opposed to GNU man)
does not support inputting a manpage on standard input. So now we pipe a
text to 'less' instead. (The information -- still extracted from the
source code POD -- still looks the same, except that bold, italic etc.
is replaced by ASCII formatting.)